Fixed some security issues
Please note that a session is shared across a browser. Hence, logging in with one user on a
tab, and another on a second tab would refresh the first page to be logged in with the second
user (Use incognito mode for logging in with a second user to overcome this).